+++ CHANGELOG.md
@@ -77,6 +77,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   during partial lookups to match the opath resolver behaviour and to avoid
   confusion by users (it is theoretically safe to fall back from a
   `SafetyViolation` during a partial lookup, but it's better to be safe here).
+- The error handling for `Root::*` operations that require splitting paths into
+  a parent directory and single basename component (such as `Root::create`) has
+  now been unified and cases like trailing `/.` and `/..` will now always
+  result in `ErrorKind::InvalidArgument`.
+- Trailing slash behaviour (i.e. where a user specifies a trailing slash in a
+  path passed to libpathrs) throughout libpathrs has been improved to better
+  match the kernel APIs (where possible) or otherwise has been made consistent
+  and intentional:
+  - `Root::create` will always error out with an `InvalidArgument` for the
+    target path unless the inode being created is an `InodeType::Directory`, in
+    which case the trailing slash will be ignored (to match the behaviour of
+    `mkdir(2)` on Linux). Hard links with a trailing slash will also produce an
+    error, as hard-links to directories are also forbidden on Unix.
+  - `Root::create_file` will always error out with an `InvalidArgument`.
+  - `Root::remove_all` and `Root::remove_dir` will ignore trailing slashes,
+    while `Root::remove_file` will always fail with `ENOTDIR`. The reason for
+    `Root::remove_all` always succeeding is that it matches the behaviour of
+    Go's `os.RemoveAll` and `rm -rf`, as well as being impractical for us to
+    determine if the target to be deleted is a directory in a race-free way.
+  - `Root::rename` matches `renameat2(2)`'s behaviour to the best of our
+    ability.
+    * Trailing slashes on the source path are only allowed if the source is
+      actually a directory (otherwise you get `ENOTDIR`).
+    * For `RENAME_EXCHANGE`, the target path may only have trailing slashes if
+      it is actually a directory (same as the source path). Otherwise, if the
+      *target* path has a trailing slash then the *source* path must be a
+      directory (otherwise you get `ENOTDIR`).
 
 ### Changed ###
 - syscalls: switch to rustix for most of our syscall wrappers to simplify how